langchain/2.langchain-and-RAG/requirements.txt
# Core LangChain
langchain==0.2.11
langchain-core==0.2.43
langchain-community==0.2.10
langchain-experimental==0.0.62
langchainhub==0.1.18

# OpenAI
langchain-openai==0.1.14
openai==1.109.1

# Vector DB & docs
chromadb==0.4.24
pypdf==4.2.0

# Env variables
python-dotenv==1.0.1

# for downloading files
wget==3.2
gradio==6.3.0
